refactor the card component to stac info card and composed sam info card
